% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPQH.R
\name{FGH}
\alias{FGH}
\alias{rndMixture}
\title{The SCIs Based on FGH Methods}
\usage{
FGH(n,p,mu,sigma,N,C2=rbind(c(-1,1,0),c(-1,0,1),c(0,-1,1)),alpha=0.05)
}
\arguments{
\item{p}{The zero probability of the mixture distribution,it has the same length to the \strong{n} params.}

\item{N}{The number of independent generated data sets.}

\item{n}{The sample size of the mixture distributions,must be an integer vector.}

\item{mu}{The mean of the non-zero samples,which after log-transformation.}

\item{sigma}{The standard deviation of the non-zero samples,which after log-transformation.}

\item{C2}{Matrix C,You can refer to the paper of Xu et al. for specific forms.}

\item{alpha}{The confidence level,it always set \emph{alpha=0.5}}
}
\value{
The method will return the Simultaneous Confidence Intervals(SCIs) and the time consuming
}
\description{
A method based on generalized pivotal quantity with order statistics(also see \code{\link{FGW}}) to construct the simultaneous confidence intervals for
Ratios of Means of Log-normal Populations with excess Zeros.
}
\details{
More information about FGH, you can read the paper: Simultaneous Confidence Intervals for Ratios of Means of Log-normal Populations with Zeros.
}
\examples{


alpha <- 0.05

p <- c(0.1,0.15,0.1)
n <- c(30,15,50)
mu <- c(0,0,0)
sigma <- c(1,1,1)
N <- 500
FGH(n,p,mu,sigma,N)

\dontrun{
p <- c(0.1,0.15,0.1,0.6)
n <- c(30,15,10,50)
mu <- c(0,0,0,0)
sigma <- c(1,1,1,1)
C2 <- rbind(c(-1,1,0,0),c(-1,0,1,0),c(-1,0,0,1),c(0,-1,1,0),c(0,-1,0,1),c(0,0,-1,1))

N <- 1000;
FGH(n,p,mu,sigma,N,C2 = C2)
}


}
\author{
Jing Xu, Xinmin Li, Hua Liang
}
